+++ ModAudio/Scripting/Data/AtlyssModule.cs
@@ -10,20 +10,20 @@ namespace Marioalexsan.ModAudio.Scripting.Data;
 public partial class AtlyssModule
 {
     [LuaMember("mainPlayer")]
-    public static PlayerProxy? MainPlayer => Player._mainPlayer != null ? new PlayerProxy(Player._mainPlayer) : null;
+    public static PlayerProxy? MainPlayer => PlayerProxy.Proxy(Player._mainPlayer);
 
     [LuaMember("actionBarManager")]
-    public static ActionBarManagerProxy ActionBarManager => new ActionBarManagerProxy(global::ActionBarManager._current);
+    public static ActionBarManagerProxy? ActionBarManager => ActionBarManagerProxy.Proxy(global::ActionBarManager._current);
 
     [LuaMember("gameWorldManager")]
-    public static GameWorldManagerProxy GameWorldManager => new GameWorldManagerProxy(global::GameWorldManager._current);
+    public static GameWorldManagerProxy? GameWorldManager => GameWorldManagerProxy.Proxy(global::GameWorldManager._current);
 
     [LuaMember("shopkeepManager")]
-    public static ShopkeepManagerProxy ShopkeepManager => new ShopkeepManagerProxy(global::ShopkeepManager._current);
+    public static ShopkeepManagerProxy? ShopkeepManager => ShopkeepManagerProxy.Proxy(global::ShopkeepManager._current);
 
     [LuaMember("mainMenuManager")]
-    public static MainMenuManagerProxy MainMenuManager => new MainMenuManagerProxy(global::MainMenuManager._current);
+    public static MainMenuManagerProxy? MainMenuManager => MainMenuManagerProxy.Proxy(global::MainMenuManager._current);
 
     [LuaMember("inGameUI")]
-    public static InGameUIProxy InGameUI => new InGameUIProxy(global::InGameUI._current);
+    public static InGameUIProxy? InGameUI => InGameUIProxy.Proxy(global::InGameUI._current);
 }