+++ ModAudio/Scripting/Proxies/MainMenuManagerProxy.cs
@@ -3,7 +3,23 @@
 namespace Marioalexsan.ModAudio.Scripting.Proxies;
 
 [LuaObject]
-public partial class MainMenuManagerProxy(MainMenuManager target)
+public partial class MainMenuManagerProxy
 {
-    public readonly MainMenuManager Target = target;
+    private MainMenuManagerProxy(MainMenuManager target)
+    {
+        Target = target;
+    }
+
+    public static MainMenuManagerProxy? Proxy(MainMenuManager? target) => target != null ? new(target) : null;
+
+    public readonly MainMenuManager Target;
+
+    [LuaMember]
+    public int _mainMenuSelection => Target._mainMenuSelection;
+
+    [LuaMember]
+    public int _multiplayerMenuSelection => Target._multiplayerMenuSelection;
+
+    [LuaMember]
+    public bool _inMultiplayerMenu => Target._inMultiplayerMenu;
 }